+++ src/components/LatestNews/LatestNews.jsx
@@ -1,11 +1,15 @@
+import { getLatestNews } from '../../api/apiNews';
+import { useFetch } from '../../helpers/Hooks/useFetch';
 import BannersList from '../BannersList/BannersList';
 import styles from './LatestNews.module.css'
 
-const LatestNews = ({banners, isLoading}) => {
+const LatestNews = () => {
+  const {data, isLoading} = useFetch(getLatestNews);
+
   return (
     <section className={styles.section} >
       <BannersList 
-        banners={banners}
+        banners={data && data.news}
         isLoading={isLoading}
       />
     </section>
