+++ src/components/NewsByFilters/NewsByFilters.jsx
@@ -1,12 +1,27 @@
-/* eslint-disable react/prop-types */
-import { TOTAL_PAGES } from '../../constants/constant'
+import { getNews } from '../../api/apiNews'
+import { PAGE_SIZE, TOTAL_PAGES } from '../../constants/constant'
+import { useDebounce } from '../../helpers/Hooks/useDebounce'
+import { useFetch } from '../../helpers/Hooks/useFetch'
+import { useFilters } from '../../helpers/Hooks/useFilters'
 import NewsFilters from '../NewsFilters/NewsFilters'
 import NewsList from '../NewsList/NewsList'
-import Pagination from '../Pagination/Pagination'
+import PaginationWrapper from '../PaginationWrapper/PaginationWrapper'
 import styles from './NewsByFilters.module.css'
 
-const NewsByFilters = ({filters, changeFilter, isLoading, news}) => {
+const NewsByFilters = () => {
+  const {filters, changeFilter} = useFilters({
+    page_number: 1, 
+    page_size: PAGE_SIZE,
+    category: null,
+    keywords: '',
+});
 
+  const debouncedKeywords = useDebounce(filters.keywords, 1500);
+
+  const {data, isLoading} = useFetch(getNews, {
+    ...filters,
+    keywords: debouncedKeywords,
+  });
 
   const handleNextPage = () => {
     if(filters.page_number < TOTAL_PAGES) {
@@ -31,26 +46,17 @@ const NewsByFilters = ({filters, changeFilter, isLoading, news}) => {
           changeFilter={changeFilter}
         />
 
-        <Pagination
+        <PaginationWrapper
+          top
+          bottom
           totalPages={TOTAL_PAGES} 
           handleNextPage={handleNextPage}
           handlePageClick={handlePageClick}
           handlePreviousPage={handlePreviousPage}
           currentPage={filters.page_number}
-        />
-
-        <NewsList
-          isLoading={isLoading} 
-          news={news} 
-        />
-
-        <Pagination
-          totalPages={TOTAL_PAGES} 
-          handleNextPage={handleNextPage}
-          handlePageClick={handlePageClick}
-          handlePreviousPage={handlePreviousPage}
-          currentPage={filters.page_number}
-        />
+        >
+          <NewsList isLoading={isLoading}  news={data?.news} />
+        </PaginationWrapper>
     </section>
   )
 }
